feat(planstate): create a dedicated plan manager
This allows sharing the plan amonst multiple stakeholder managers.
