+++ internals/daemon/api_exec.go
@@ -64,7 +64,7 @@ func v1PostExec(c *Command, req *http.Request, _ *UserState) Response {
 		return statusBadRequest("cannot find executable %q", payload.Command[0])
 	}
 
-	p, err := c.d.overlord.ServiceManager().Plan()
+	p, err := c.d.overlord.PlanManager().Plan()
 	if err != nil {
 		return statusBadRequest("%v", err)
 	}
