+++ internals/overlord/servstate/manager.go
@@ -12,6 +12,7 @@ import (
 	"github.com/canonical/pebble/internals/logger"
 	"github.com/canonical/pebble/internals/overlord/restart"
 	"github.com/canonical/pebble/internals/overlord/state"
+	"github.com/canonical/pebble/internals/overlord/planstate"
 	"github.com/canonical/pebble/internals/plan"
 	"github.com/canonical/pebble/internals/reaper"
 	"github.com/canonical/pebble/internals/servicelog"
@@ -20,11 +21,8 @@ import (
 type ServiceManager struct {
 	state     *state.State
 	runner    *state.TaskRunner
-	pebbleDir string
-
-	planLock     sync.Mutex
-	plan         *plan.Plan
-	planHandlers []PlanFunc
+	plan      *plan.Plan
+	planMgr   *planstate.PlanManager
 
 	servicesLock sync.Mutex
 	services     map[string]*serviceData
@@ -42,28 +40,15 @@ type LogManager interface {
 	ServiceStarted(service *plan.Service, logs *servicelog.RingBuffer)
 }
 
-// PlanFunc is the type of function used by NotifyPlanChanged.
-type PlanFunc func(p *plan.Plan)
-
 type Restarter interface {
 	HandleRestart(t restart.RestartType)
 }
 
-// LabelExists is the error returned by AppendLayer when a layer with that
-// label already exists.
-type LabelExists struct {
-	Label string
-}
-
-func (e *LabelExists) Error() string {
-	return fmt.Sprintf("layer %q already exists", e.Label)
-}
-
-func NewManager(s *state.State, runner *state.TaskRunner, pebbleDir string, serviceOutput io.Writer, restarter Restarter, logMgr LogManager) (*ServiceManager, error) {
+func NewManager(s *state.State, runner *state.TaskRunner, planMgr *planstate.PlanManager, serviceOutput io.Writer, restarter Restarter, logMgr LogManager) (*ServiceManager, error) {
 	manager := &ServiceManager{
 		state:         s,
 		runner:        runner,
-		pebbleDir:     pebbleDir,
+		planMgr:       planMgr,
 		services:      make(map[string]*serviceData),
 		serviceOutput: serviceOutput,
 		restarter:     restarter,
@@ -82,6 +67,13 @@ func NewManager(s *state.State, runner *state.TaskRunner, pebbleDir string, serv
 	return manager, nil
 }
 
+// planUpdated informs the service manager that the plan has been updated.
+func (m *ServiceManager) PlanChanged(plan *plan.Plan) {
+	m.servicesLock.Lock()
+	defer m.servicesLock.Unlock()
+	m.plan = plan
+}
+
 // Stop implements overlord.StateStopper and stops background functions.
 func (m *ServiceManager) Stop() {
 	err := reaper.Stop()
@@ -95,153 +87,6 @@ func (m *ServiceManager) Stop() {
 	for name := range m.services {
 		m.removeServiceInternal(name)
 	}
-}
-
-// NotifyPlanChanged adds f to the list of functions that are called whenever
-// the plan is updated.
-func (m *ServiceManager) NotifyPlanChanged(f PlanFunc) {
-	m.planHandlers = append(m.planHandlers, f)
-}
-
-func (m *ServiceManager) updatePlan(p *plan.Plan) {
-	m.plan = p
-	for _, f := range m.planHandlers {
-		f(p)
-	}
-}
-
-func (m *ServiceManager) reloadPlan() error {
-	p, err := plan.ReadDir(m.pebbleDir)
-	if err != nil {
-		return err
-	}
-	m.updatePlan(p)
-	return nil
-}
-
-// Plan returns the configuration plan.
-func (m *ServiceManager) Plan() (*plan.Plan, error) {
-	releasePlan, err := m.acquirePlan()
-	if err != nil {
-		return nil, err
-	}
-	defer releasePlan()
-	return m.plan, nil
-}
-
-// AppendLayer appends the given layer to the plan's layers and updates the
-// layer.Order field to the new order. If a layer with layer.Label already
-// exists, return an error of type *LabelExists.
-func (m *ServiceManager) AppendLayer(layer *plan.Layer) error {
-	releasePlan, err := m.acquirePlan()
-	if err != nil {
-		return err
-	}
-	defer releasePlan()
-
-	index, _ := findLayer(m.plan.Layers, layer.Label)
-	if index >= 0 {
-		return &LabelExists{Label: layer.Label}
-	}
-
-	return m.appendLayer(layer)
-}
-
-func (m *ServiceManager) appendLayer(layer *plan.Layer) error {
-	newOrder := 1
-	if len(m.plan.Layers) > 0 {
-		last := m.plan.Layers[len(m.plan.Layers)-1]
-		newOrder = last.Order + 1
-	}
-
-	newLayers := append(m.plan.Layers, layer)
-	err := m.updatePlanLayers(newLayers)
-	if err != nil {
-		return err
-	}
-	layer.Order = newOrder
-	return nil
-}
-
-func (m *ServiceManager) updatePlanLayers(layers []*plan.Layer) error {
-	combined, err := plan.CombineLayers(layers...)
-	if err != nil {
-		return err
-	}
-	p := &plan.Plan{
-		Layers:     layers,
-		Services:   combined.Services,
-		Checks:     combined.Checks,
-		LogTargets: combined.LogTargets,
-	}
-	m.updatePlan(p)
-	return nil
-}
-
-// findLayer returns the index (in layers) of the layer with the given label,
-// or returns -1, nil if there's no layer with that label.
-func findLayer(layers []*plan.Layer, label string) (int, *plan.Layer) {
-	for i, layer := range layers {
-		if layer.Label == label {
-			return i, layer
-		}
-	}
-	return -1, nil
-}
-
-// CombineLayer combines the given layer with an existing layer that has the
-// same label. If no existing layer has the label, append a new one. In either
-// case, update the layer.Order field to the new order.
-func (m *ServiceManager) CombineLayer(layer *plan.Layer) error {
-	releasePlan, err := m.acquirePlan()
-	if err != nil {
-		return err
-	}
-	defer releasePlan()
-
-	index, found := findLayer(m.plan.Layers, layer.Label)
-	if index < 0 {
-		// No layer found with this label, append new one.
-		return m.appendLayer(layer)
-	}
-
-	// Layer found with this label, combine into that one.
-	combined, err := plan.CombineLayers(found, layer)
-	if err != nil {
-		return err
-	}
-	combined.Order = found.Order
-	combined.Label = found.Label
-
-	// Insert combined layer back into plan's layers list.
-	newLayers := make([]*plan.Layer, len(m.plan.Layers))
-	copy(newLayers, m.plan.Layers)
-	newLayers[index] = combined
-	err = m.updatePlanLayers(newLayers)
-	if err != nil {
-		return err
-	}
-	layer.Order = found.Order
-	return nil
-}
-
-func (m *ServiceManager) acquirePlan() (release func(), err error) {
-	m.planLock.Lock()
-	if m.plan == nil {
-		err := m.reloadPlan()
-		if err != nil {
-			m.planLock.Unlock()
-			return nil, err
-		}
-	}
-	released := false
-	release = func() {
-		if !released {
-			released = true
-			m.planLock.Unlock()
-		}
-	}
-	return release, nil
 }
 
 // Ensure implements StateManager.Ensure.
@@ -275,12 +120,6 @@ const (
 // Services returns the list of configured services and their status, sorted
 // by service name. Filter by the specified service names if provided.
 func (m *ServiceManager) Services(names []string) ([]*ServiceInfo, error) {
-	releasePlan, err := m.acquirePlan()
-	if err != nil {
-		return nil, err
-	}
-	defer releasePlan()
-
 	m.servicesLock.Lock()
 	defer m.servicesLock.Unlock()
 
@@ -355,12 +194,6 @@ func stateToStatus(state serviceState) ServiceStatus {
 // DefaultServiceNames returns the name of the services set to start
 // by default.
 func (m *ServiceManager) DefaultServiceNames() ([]string, error) {
-	releasePlan, err := m.acquirePlan()
-	if err != nil {
-		return nil, err
-	}
-	defer releasePlan()
-
 	var names []string
 	for name, service := range m.plan.Services {
 		if service.Startup == plan.StartupEnabled {
@@ -374,24 +207,12 @@ func (m *ServiceManager) DefaultServiceNames() ([]string, error) {
 // StartOrder returns the provided services, together with any required
 // dependencies, in the proper order for starting them all up.
 func (m *ServiceManager) StartOrder(services []string) ([]string, error) {
-	releasePlan, err := m.acquirePlan()
-	if err != nil {
-		return nil, err
-	}
-	defer releasePlan()
-
 	return m.plan.StartOrder(services)
 }
 
 // StopOrder returns the provided services, together with any dependants,
 // in the proper order for stopping them all.
 func (m *ServiceManager) StopOrder(services []string) ([]string, error) {
-	releasePlan, err := m.acquirePlan()
-	if err != nil {
-		return nil, err
-	}
-	defer releasePlan()
-
 	return m.plan.StopOrder(services)
 }
 
@@ -399,12 +220,6 @@ func (m *ServiceManager) StopOrder(services []string) ([]string, error) {
 // return tail iterators; if last is zero or positive, return head iterators
 // going back last elements. Each iterator must be closed via the Close method.
 func (m *ServiceManager) ServiceLogs(services []string, last int) (map[string]servicelog.Iterator, error) {
-	releasePlan, err := m.acquirePlan()
-	if err != nil {
-		return nil, err
-	}
-	defer releasePlan()
-
 	requested := make(map[string]bool, len(services))
 	for _, name := range services {
 		requested[name] = true
@@ -434,12 +249,6 @@ func (m *ServiceManager) ServiceLogs(services []string, last int) (map[string]se
 // Replan returns a list of services to stop and services to start because
 // their plans had changed between when they started and this call.
 func (m *ServiceManager) Replan() ([]string, []string, error) {
-	releasePlan, err := m.acquirePlan()
-	if err != nil {
-		return nil, nil, err
-	}
-	defer releasePlan()
-
 	m.servicesLock.Lock()
 	defer m.servicesLock.Unlock()
 
@@ -463,7 +272,7 @@ func (m *ServiceManager) Replan() ([]string, []string, error) {
 		}
 	}
 
-	stop, err = m.plan.StopOrder(stop)
+	stop, err := m.plan.StopOrder(stop)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -524,12 +333,6 @@ func (m *ServiceManager) CheckFailed(name string) {
 // to their respective services. It adds a new layer in the plan, the layer
 // consisting of services with commands having their arguments changed.
 func (m *ServiceManager) SetServiceArgs(serviceArgs map[string][]string) error {
-	releasePlan, err := m.acquirePlan()
-	if err != nil {
-		return err
-	}
-	defer releasePlan()
-
 	newLayer := &plan.Layer{
 		// TODO: Consider making any labels starting with
 		// the "pebble-" prefix reserved.
@@ -552,7 +355,7 @@ func (m *ServiceManager) SetServiceArgs(serviceArgs map[string][]string) error {
 		}
 	}
 
-	return m.appendLayer(newLayer)
+	return m.planMgr.AppendLayer(newLayer)
 }
 
 // servicesToStop is used during service manager shutdown to cleanly terminate
@@ -564,12 +367,6 @@ func (m *ServiceManager) SetServiceArgs(serviceArgs map[string][]string) error {
 // exits), it would generate a runtime error as the reaper would already be dead.
 // This function returns a slice of service names to stop, in dependency order.
 func servicesToStop(m *ServiceManager) ([]string, error) {
-	releasePlan, err := m.acquirePlan()
-	if err != nil {
-		return nil, err
-	}
-	defer releasePlan()
-
 	// Get all service names in plan.
 	var services []string
 	for name := range m.plan.Services {
